backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
google-generativeai==0.5.0
pydantic==2.7.1
python-dotenv==1.0.1